requirements.txt
selenium
llama-cpp-python
transformers
langchain
beautifulsoup4
duckduckgo-search
chromadb
pathvalidate
pydantic
docstring-parser
requests